refactor(gui): move ExhibitionEntityEditorScreen to main gui package
